Complete TS Type Generation!
Wow, it fully works! Now I just have to format this output a bit more so each map item is it's own interface that can be used on it's own. I might use IDE refactoring to do that more efficiently. (Move to it's own type; not sure if that's a thing or not for types, like variables are)