Block State Property Sorting
I think these appear to be sorted alphabetically already, but doing this only helps to ensure less things to diff when working with the output from this script, say if the keys do change order for one reason or another. I'm glad Bedrock does sort it's keys too, it really does help ensure it's easy to read through NBT with your eyes a little easier.

Yeah so the output for this file at least, doesn't change at all with this additional sorter. This is to account for files that aren't sorted by chance for some reason. I'm going to bring some of these concepts back to the definition writer for Region-Types too, having sorting and things like that actually would help a bunch. Same with the raw values typings thing, which this version is mainly forked because of.