Automatic Interface Refactor Setup
When adding the types output from this into Region-Types itself, I had to manually extract each interface from the type map into their own declarations, and this took a little bit extra work. Since these objects are already something tangibly iterable in terms of JS land, I might as well add the manipulation for that into the string builder itself. Then I don't have to do those refactorings myself with each update to the source JSON file. I was nearly there with having the types declared on their own in the types file itself, but I had to come up with some way to rename the declarations to PascalCase, and I didn't want to have to do that every single time as well. The extraction part wasn't very hard because of cursor multi-select, it was just that more things adding up made it seem like abstracting those out as well only made more sense in the long run. Fun exercise as well!

Rediscovered Crossy Road, forgot how much I really loved that game! They made it work with a controller nowadays too, it's a really fun experience on my iPad! I think it's one of my favorite, solid apps come to think of it. I really enjoy all of the details they put into the game, as well as the UI and UX are extraordinary, at least for the standards of other mobile games. Minecraft is losing on that front, haha. It even runs on macOS by way of Apple Silicon, as the iPad app! Extra awesome, even the keyboard works with it as well. Great details there. I wish Bedrock was just available on macOS already, omg. How long is it going to take?

https://stackoverflow.com/questions/44082153/javascript-method-for-changing-snake-case-to-pascalcase