Field Type Persistence Debug
This was heavily improperly calculated previously, it was lossily converting some other keys in each block state to types they weren't meant to, and this made a lot of the values typed with the wrong primitive, and sometimes even coalesced the values to then be wrong too. This fixes the issues with the wall values, which is where I noticed the missing string tag values. It was accidentally converting them to Byte tag values because neighboring keys were using that. Now each property properly gets it's own type, which is what I intended to do altogether.